Support https nodes

+++ backend/util/iri.util.js
@@ -3,6 +3,7 @@ const fs = require('fs');
 class IriUtil {
 
   constructor() {
+    this.protocol = null;
     this.iriIp = null;
     this.iriPort = null;
     this.iriFileLocation = null;
@@ -10,7 +11,7 @@ class IriUtil {
 
   createIriRequest(command, ip = this.iriIp) {
     return {
-      url: `http://${ip}:${this.iriPort}`,
+      url: `${this.protocol}://${ip}:${this.iriPort}`,
       data: {command},
       method: 'post',
       headers: {
@@ -40,14 +41,16 @@ class IriUtil {
 
   // Only works if at least one neighbor is remaining
   removeNeighborToIriConfig(fullAddress) {
-    fs.readFile(this.iriFileLocation, 'utf-8', (err, data) => {
-      if (data.includes(fullAddress)) {
-        const withRemovedNeighbor = data.replace(`${fullAddress} `, '');
-        fs.writeFile(this.iriFileLocation, withRemovedNeighbor, (err) => {
-          if (err) throw err;
-        });
-      }
-    });
+    if (fs.existsSync(this.iriFileLocation)) {
+      fs.readFile(this.iriFileLocation, 'utf-8', (err, data) => {
+        if (data.includes(fullAddress)) {
+          const withRemovedNeighbor = data.replace(`${fullAddress} `, '');
+          fs.writeFile(this.iriFileLocation, withRemovedNeighbor, (err) => {
+            if (err) throw err;
+          });
+        }
+      });
+    }
   }
 
 }
